Improve naming and structure of HbmXmlComplexFileTest (now HbmXmlComponentTest)
"Complex" could mean anything; at least "Component" tells us we're
testing the "component" feature of hbm.xml.

Also, yes, I agree "simple" in "hbm-simple.xml" is not great either.
I just used this to mean "there's nothing special about it, we simply
want to test the basics of HBM file support".
