+++ extensions/hibernate-orm/deployment/src/test/resources/META-INF/hbm-component.xml
@@ -3,11 +3,11 @@
         "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
         "http://www.hibernate.org/dtd/hibernate-mapping-3.0.dtd">
 <hibernate-mapping>
-    <class name="io.quarkus.hibernate.orm.xml.hbm.NonAnnotatedComplexEntity">
+    <class name="io.quarkus.hibernate.orm.xml.hbm.NonAnnotatedComponentUsingEntity">
         <id name="id">
             <generator class="identity"/>
         </id>
-        <component name="value" class="io.quarkus.hibernate.orm.xml.hbm.NonAnnotatedComponentEntity" lazy="false">
+        <component name="value" class="io.quarkus.hibernate.orm.xml.hbm.NonAnnotatedComponent" lazy="false">
             <property name="name">
                 <column name="thename"/>
             </property>
