--- extensions/hibernate-orm/deployment/src/test/resources/application-mapping-files-my-complex-hbm-xml.properties
@@ -1,7 +1,0 @@
-quarkus.datasource.db-kind=h2
-quarkus.datasource.jdbc.url=jdbc:h2:mem:test
-
-quarkus.hibernate-orm.dialect=org.hibernate.dialect.H2Dialect
-#quarkus.hibernate-orm.log.sql=true
-quarkus.hibernate-orm.database.generation=drop-and-create
-quarkus.hibernate-orm.mapping-files=my-complex-hbm.xml
